chore: cleanup oracles hook
