+++ packages/apps/human-app/frontend/src/modules/worker/jobs/hooks/use-job-types-oracles-table.tsx
@@ -1,18 +1,17 @@
-/* eslint-disable camelcase -- ... */
 import { create } from 'zustand';
 
 export interface JobsTypesOraclesFilterStore {
-  selected_job_types: string[];
+  selectedJobTypes: string[];
   selectJobType: (jobType: string[]) => void;
 }
 
 export const useJobsTypesOraclesFilterStore =
   create<JobsTypesOraclesFilterStore>((set) => ({
-    selected_job_types: [],
+    selectedJobTypes: [],
     selectJobType: (jobTypes: string[]) => {
       set((state) => ({
         ...state,
-        selected_job_types: jobTypes,
+        selectedJobTypes: jobTypes,
       }));
     },
   }));
